+++ tests/testthat/test-range.R
@@ -44,5 +44,5 @@ test_that("factor discrete ranges stay in order", {
   expect_equal(discrete_range(f, f), letters[3:1])
   expect_equal(discrete_range(f, "c"), letters[3:1])
   expect_equal(discrete_range(f, c("a", "b", "c")), letters[3:1])
-  expect_equal(discrete_range(f, c("a", "b", "c", NA), na.rm = FALSE), letters[3:1])
+  expect_equal(discrete_range(f, c("a", "b", "c", NA), na.rm = FALSE), c(letters[3:1], NA))
 })